test: clean up test suite
- Move shared fixtures to conftest.py
- Remove redundant inline imports of bpy in test files
- Remove unused fixture and empty placeholder tests
- Remove descriptive path comments
